Fixed 64-bit portability warnings in windows sys.
Note: these places were buggy in 64-bit, but I guess there are never executed during normal operation.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7496 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ sys/win32/win_main.cpp
@@ -761,7 +761,7 @@ uintptr_t Sys_DLL_Load(const char *dllName) {
 
 		if ( idStr::IcmpPath( dllPath, loadedPath ) ) {
 			Sys_Printf( "ERROR: LoadLibrary '%s' wants to load '%s'\n", dllPath, loadedPath );
-			Sys_DLL_Unload( (int)libHandle );
+			Sys_DLL_Unload( (uintptr_t)libHandle );
 			return 0;
 		}
 	}
